lab3_py/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
minio==7.2.5
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1